+++ src/bk-user/bkuser/apis/open_v3/views/user.py
@@ -73,7 +73,7 @@ class TenantUserDisplayInfoListApi(OpenApiCommonMixin, generics.ListAPIView):
         return TenantUser.objects.filter(
             id__in=data["bk_usernames"],
             tenant_id=self.tenant_id,
-            data_source_id=self.real_data_source_id,
+            data_source_id__in=self.real_data_source_ids,
         ).select_related("data_source_user")
 
     def get_serializer_context(self):
@@ -104,7 +104,7 @@ class TenantUserRetrieveApi(OpenApiCommonMixin, generics.RetrieveAPIView):
 
     def get_queryset(self):
         return TenantUser.objects.filter(
-            tenant_id=self.tenant_id, data_source_id=self.real_data_source_id
+            tenant_id=self.tenant_id, data_source_id__in=self.real_data_source_ids
         ).select_related("data_source_user", "data_source")
 
     def get_serializer_context(self):
@@ -140,7 +140,7 @@ class TenantUserDepartmentListApi(OpenApiCommonMixin, generics.ListAPIView):
         data = slz.validated_data
 
         tenant_user = get_object_or_404(
-            TenantUser.objects.filter(tenant_id=self.tenant_id, data_source_id=self.real_data_source_id),
+            TenantUser.objects.filter(tenant_id=self.tenant_id, data_source_id__in=self.real_data_source_ids),
             id=kwargs["id"],
         )
 
@@ -216,7 +216,7 @@ class TenantUserLeaderListApi(OpenApiCommonMixin, generics.ListAPIView):
 
     def get_queryset(self) -> QuerySet[TenantUser]:
         tenant_user = get_object_or_404(
-            TenantUser.objects.filter(tenant_id=self.tenant_id, data_source_id=self.real_data_source_id),
+            TenantUser.objects.filter(tenant_id=self.tenant_id, data_source_id__in=self.real_data_source_ids),
             id=self.kwargs["id"],
         )
 
@@ -260,7 +260,7 @@ class TenantUserListApi(OpenApiCommonMixin, generics.ListAPIView):
     def get_queryset(self) -> QuerySet[TenantUser]:
         return (
             TenantUser.objects.select_related("data_source_user")
-            .filter(tenant_id=self.tenant_id, data_source_id=self.real_data_source_id)
+            .filter(tenant_id=self.tenant_id, data_source_id__in=self.real_data_source_ids)
             .order_by("id")
         )
 
@@ -297,7 +297,7 @@ class TenantUserSensitiveInfoListApi(OpenApiCommonMixin, generics.ListAPIView):
         data = slz.validated_data
 
         return TenantUser.objects.filter(
-            id__in=data["bk_usernames"], tenant_id=self.tenant_id, data_source_id=self.real_data_source_id
+            id__in=data["bk_usernames"], tenant_id=self.tenant_id, data_source_id__in=self.real_data_source_ids
         ).select_related("data_source_user")
 
     @swagger_auto_schema(
@@ -326,7 +326,7 @@ class TenantUserLookupApi(OpenApiCommonMixin, generics.ListAPIView):
 
         filter_args = {
             "tenant_id": self.tenant_id,
-            "data_source_id": self.real_data_source_id,
+            "data_source_id__in": self.real_data_source_ids,
         }
 
         if data["lookup_field"] == UserLookupFieldEnum.LOGIN_NAME:
