+++ src/bk-user/bkuser/apis/open_web/serializers/users.py
@@ -24,7 +24,7 @@ from rest_framework.exceptions import ValidationError
 from bkuser.apps.data_source.constants import DataSourceTypeEnum
 from bkuser.apps.tenant.constants import TenantUserStatus
 from bkuser.apps.tenant.models import TenantUser
-from bkuser.biz.tenant import TenantUserDisplayNameHandler, get_tenant_user_login_name
+from bkuser.biz.tenant import TenantUserDisplayNameHandler, TenantUserHandler
 from bkuser.common.constants import BkLanguageEnum
 from bkuser.common.serializers import StringArrayField
 
@@ -35,7 +35,7 @@ class TenantUserDisplayInfoRetrieveOutputSLZ(serializers.Serializer):
     display_name = serializers.SerializerMethodField(help_text="用户展示名称")
 
     def get_login_name(self, obj: TenantUser) -> str:
-        return get_tenant_user_login_name(self.context["tenant_id"], obj)
+        return TenantUserHandler.get_login_name(obj, self.context["tenant_id"])
 
     def get_display_name(self, obj: TenantUser) -> str:
         return TenantUserDisplayNameHandler.generate_tenant_user_display_name(obj)
@@ -55,7 +55,7 @@ class TenantUserDisplayInfoListOutputSLZ(serializers.Serializer):
     display_name = serializers.SerializerMethodField(help_text="用户展示名称")
 
     def get_login_name(self, obj: TenantUser) -> str:
-        return get_tenant_user_login_name(self.context["tenant_id"], obj)
+        return TenantUserHandler.get_login_name(obj, self.context["tenant_id"])
 
     def get_display_name(self, obj: TenantUser) -> str:
         return self.context["display_name_map"][obj.id]
@@ -89,7 +89,7 @@ class TenantUserSearchOutputSLZ(serializers.Serializer):
     organization_paths = serializers.SerializerMethodField(help_text="用户所属部门路径")
 
     def get_login_name(self, obj: TenantUser) -> str:
-        return get_tenant_user_login_name(self.context["tenant_id"], obj)
+        return TenantUserHandler.get_login_name(obj, self.context["tenant_id"])
 
     def get_display_name(self, obj: TenantUser) -> str:
         return self.context["display_name_map"][obj.id]
@@ -146,7 +146,7 @@ class TenantUserLookupOutputSLZ(serializers.Serializer):
     organization_paths = serializers.SerializerMethodField(help_text="用户所属部门路径")
 
     def get_login_name(self, obj: TenantUser) -> str:
-        return get_tenant_user_login_name(self.context["tenant_id"], obj)
+        return TenantUserHandler.get_login_name(obj, self.context["tenant_id"])
 
     def get_display_name(self, obj: TenantUser) -> str:
         return self.context["display_name_map"][obj.id]
