+++ src/bk-user/bkuser/apis/web/organization/serializers/tenants.py
@@ -51,18 +51,18 @@ class TenantListOutputSLZ(serializers.Serializer):
 
 
 class TenantRetrieveOutputSLZ(TenantListOutputSLZ):
-    data_source = serializers.SerializerMethodField(help_text="实名用户数据源信息")
+    data_sources = serializers.SerializerMethodField(help_text="实名用户数据源信息列表")
 
     class Meta:
         ref_name = "organization.TenantRetrieveOutputSLZ"
 
-    @swagger_serializer_method(serializer_or_field=TenantDataSourceSLZ())
-    def get_data_source(self, obj: Tenant) -> Dict[str, Any] | None:
-        data_source = DataSource.objects.filter(owner_tenant_id=obj.id, type=DataSourceTypeEnum.REAL).first()
-        if not data_source:
-            return None
+    @swagger_serializer_method(serializer_or_field=TenantDataSourceSLZ(many=True))
+    def get_data_sources(self, obj: Tenant) -> list[Dict[str, Any]]:
+        data_sources = list(DataSource.objects.filter(owner_tenant_id=obj.id, type=DataSourceTypeEnum.REAL))
+        if not data_sources:
+            return []
 
-        return TenantDataSourceSLZ(data_source).data
+        return TenantDataSourceSLZ(data_sources, many=True).data
 
 
 class RequiredTenantUserFieldOutputSLZ(serializers.Serializer):
