+++ src/bk-user/bkuser/apis/web/organization/views/mixins.py
@@ -14,6 +14,8 @@
 #
 # We undertake not to change the open source license (MIT license) applicable
 # to the current version of the project delivered to anyone in the future.
+from typing import List
+
 from django.utils.translation import gettext_lazy as _
 
 from bkuser.apis.web.mixins import CurrentUserTenantMixin
@@ -25,18 +27,19 @@ from bkuser.common.error_codes import error_codes
 class CurrentUserTenantDataSourceMixin(CurrentUserTenantMixin):
     """获取当前用户所在租户指定条件数据源"""
 
-    def get_current_tenant_real_data_source(self) -> DataSource:
+    def get_current_tenant_real_data_sources(self) -> List[DataSource]:
         data_source = DataSource.objects.filter(
             owner_tenant_id=self.get_current_tenant_id(), type=DataSourceTypeEnum.REAL
-        ).first()
+        )
         if not data_source:
             raise error_codes.DATA_SOURCE_NOT_EXIST.f(_("当前租户不存在实名用户数据源"))
 
-        return data_source
+        return list(data_source)
 
     def get_current_tenant_local_real_data_source(self) -> DataSource:
-        real_data_source = self.get_current_tenant_real_data_source()
-        if not real_data_source.is_local:
+        real_ds_list = self.get_current_tenant_real_data_sources()
+        real_data_source = next((ds for ds in real_ds_list if ds.is_local), None)
+        if not real_data_source:
             raise error_codes.DATA_SOURCE_NOT_EXIST.f(_("当前租户不存在本地实名用户数据源"))
 
         return real_data_source
