+++ src/bk-user/bkuser/biz/tenant/display_name.py
@@ -1,0 +1,232 @@
+# -*- coding: utf-8 -*-
+# TencentBlueKing is pleased to support the open source community by making
+# 蓝鲸智云 - 用户管理 (bk-user) available.
+# Copyright (C) 2017 Tencent. All rights reserved.
+# Licensed under the MIT License (the "License"); you may not use this file except
+# in compliance with the License. You may obtain a copy of the License at
+#
+#     http://opensource.org/licenses/MIT
+#
+# Unless required by applicable law or agreed to in writing, software distributed under
+# the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,
+# either express or implied. See the License for the specific language governing permissions and
+# limitations under the License.
+#
+# We undertake not to change the open source license (MIT license) applicable
+# to the current version of the project delivered to anyone in the future.
+import logging
+import operator
+from functools import reduce
+from typing import Dict, List
+
+from django.contrib.auth import get_user_model
+from django.db.models import Q
+
+from bkuser.apps.data_source.models import DataSource, DataSourceUser
+from bkuser.apps.tenant.constants import (
+    DISPLAY_NAME_EXPRESSION_FIELD_PATTERN,
+)
+from bkuser.apps.tenant.display_name_cache import get_display_name_config
+from bkuser.apps.tenant.models import (
+    TenantUser,
+    TenantUserCustomField,
+    TenantUserDisplayNameExpressionConfig,
+    UserBuiltinField,
+)
+
+logger = logging.getLogger(__name__)
+
+
+class TenantUserDisplayNameHandler:
+    @staticmethod
+    def generate_tenant_user_display_name(user: TenantUser) -> str:
+        """生成租户用户展示名"""
+        config = get_display_name_config(user.tenant_id, user.data_source_id)
+        return TenantUserDisplayNameHandler.render_display_name(user, config)
+
+    @staticmethod
+    def batch_generate_tenant_user_display_name(users: List[TenantUser]) -> Dict[str, str]:
+        """
+        批量生成租户用户展示名
+
+        注意：调用时需要提前连表查询 `data_source_user`，否则会导致 N+1 问题
+        """
+
+        if not users:
+            return {}
+
+        # TODO：后续加入缓存逻辑
+        # 1. 先获取 display_name 缓存
+        # 2. 对于缓存不存在的用户，则执行 batch_render_display_name 方法进行渲染
+        # 3. 将渲染后的结果保存到缓存中
+
+        return TenantUserDisplayNameHandler.batch_render_display_name(users)
+
+    @staticmethod
+    def get_tenant_user_display_name_map_by_ids(tenant_user_ids: List[str]) -> Dict[str, str]:
+        """
+        根据指定的租户用户 ID 列表，获取对应的展示用名称列表
+
+        :return: {user_id: user_display_name}
+        """
+        # 1. 尝试从 TenantUser 表根据表达式渲染出展示用名称（使用批量处理）
+        users = TenantUser.objects.select_related("data_source_user").filter(id__in=tenant_user_ids)
+        display_name_map = TenantUserDisplayNameHandler.batch_generate_tenant_user_display_name(users)
+
+        # 2. 针对可能出现的 TenantUser 中被删除的 user_id，尝试从 User 表获取展示用名称（登录过就有记录）
+        if not_exists_user_ids := set(tenant_user_ids) - set(display_name_map.keys()):
+            logger.warning(
+                "tenant user ids: %s not exists in TenantUser model, try find display name in User Model",
+                not_exists_user_ids,
+            )
+            UserModel = get_user_model()  # noqa: N806
+            for user in UserModel.objects.filter(username__in=not_exists_user_ids):
+                # FIXME (nan) get_property 有 N+1 的风险，需要处理
+                display_name_map[user.username] = user.get_property("display_name") or user.username
+
+        # 3. 前两种方式都失效，那就给啥 user_id 就返回啥，避免调用的地方还需要处理
+        if not_exists_user_ids := set(tenant_user_ids) - set(display_name_map.keys()):
+            display_name_map.update({user_id: user_id for user_id in not_exists_user_ids})
+
+        return display_name_map
+
+    @staticmethod
+    def parse_display_name_expression(tenant_id: str, expression: str) -> Dict[str, List[str]]:
+        """解析展示名表达式，返回格式为 {builtin: [内置字段列表], custom: [自定义字段列表], extra: [额外字段列表]}"""
+        fields = DISPLAY_NAME_EXPRESSION_FIELD_PATTERN.findall(expression)
+
+        # TODO: 后续需要过滤敏感字段，敏感字段不支持展示
+        builtin_field_names = {field.name for field in UserBuiltinField.objects.all()}
+
+        custom_field_names = {field.name for field in TenantUserCustomField.objects.filter(tenant_id=tenant_id)}
+
+        # TODO: 后续支持 extra 字段
+        # 集合运算 & 求交集，比遍历判断更高效
+        return {
+            "builtin": list(set(fields) & builtin_field_names),
+            "custom": list(set(fields) & custom_field_names),
+            "extra": [],
+        }
+
+    @staticmethod
+    def render_display_name(user: TenantUser, config: TenantUserDisplayNameExpressionConfig) -> str:
+        """渲染用户展示名"""
+
+        # 获取各类字段值
+        builtin_values = TenantUserDisplayNameHandler._get_builtin_field_values(user, config.builtin_fields)
+        custom_values = TenantUserDisplayNameHandler._get_custom_field_values(user, config.custom_fields)
+
+        field_value_map = {}
+        field_value_map.update(builtin_values)
+        field_value_map.update(custom_values)
+
+        # 使用正则表达式替换表达式中的字段为对应值
+        # match 是正则表达式从表达式（expression）中匹配到的对象，group(1) 是匹配到的第一个分组，即字段名
+        # 如果字段名在 field_value_map 中存在，则使用 field_value_map 中的值替换，否则使用 "-"
+        return DISPLAY_NAME_EXPRESSION_FIELD_PATTERN.sub(
+            lambda match: field_value_map.get(match.group(1), "-"), config.expression
+        )
+
+    @staticmethod
+    def batch_render_display_name(
+        users: List[TenantUser],
+    ) -> Dict[str, str]:
+        """批量渲染用户展示用名称"""
+        if not users:
+            return {}
+
+        user_display_name_map: Dict[str, str] = {}
+
+        # 遍历所有用户，渲染 display_name
+        for user in users:
+            field_value_map = {}
+            config = get_display_name_config(user.tenant_id, user.data_source_id)
+
+            builtin_values = TenantUserDisplayNameHandler._get_builtin_field_values(user, config.builtin_fields)
+            custom_values = TenantUserDisplayNameHandler._get_custom_field_values(user, config.custom_fields)
+
+            field_value_map.update(builtin_values)
+            field_value_map.update(custom_values)
+
+            # 使用表达式模板渲染展示名
+            user_display_name_map[user.id] = DISPLAY_NAME_EXPRESSION_FIELD_PATTERN.sub(
+                lambda match, field_value_map=field_value_map: field_value_map.get(match.group(1), "-"),  # type: ignore
+                config.expression,
+            )
+
+        return user_display_name_map
+
+    @staticmethod
+    def _get_builtin_field_values(user: TenantUser, builtin_fields: List[str]) -> Dict[str, str]:
+        """处理内置字段"""
+        # TODO: 内建字段后续可能也存在协同字段映射的情况，需要处理
+        # 联系方式字段映射
+        field_value_map = {}
+
+        # 联系信息字段映射
+        contact_info = {
+            "email": user.email,
+            "phone": user.phone_info[0],
+            "phone_country_code": user.phone_info[1],
+        }
+
+        for field in builtin_fields:
+            if field in contact_info:
+                field_value_map[field] = contact_info[field] or "-"
+            else:
+                field_value_map[field] = getattr(user.data_source_user, field)
+        return field_value_map
+
+    @staticmethod
+    def _get_custom_field_values(user: TenantUser, custom_fields: List[str]) -> Dict[str, str]:
+        """处理自定义字段"""
+        return {field: str(user.data_source_user.extras.get(field, "-")) for field in custom_fields}
+
+    @staticmethod
+    def build_display_name_search_queries(tenant_id: str, keyword: str) -> Q:
+        """根据不同字段类型构建展示名搜索查询条件"""
+        config = get_display_name_config(tenant_id)
+
+        # 构建内置字段查询条件
+        builtin_queries = TenantUserDisplayNameHandler._build_builtin_field_queries(config.builtin_fields, keyword)
+        # 为什么不构建自定义字段查询条件？
+        # 因为自定义字段存储在 extra 字段（JsonField）中，进行模糊搜索非常消耗资源，JSON 数据需要解析与字符串匹配
+
+        if not builtin_queries:
+            return Q()
+
+        field_queries = reduce(operator.or_, builtin_queries)
+        return field_queries & Q(data_source__owner_tenant_id=tenant_id)
+
+    @staticmethod
+    def _build_builtin_field_queries(builtin_fields: List[str], keyword: str) -> List[Q]:
+        """构建内置字段查询条件"""
+        inherit_flag_mapping = {"phone_country_code": "phone", "phone": "phone", "email": "email"}
+        queries = []
+
+        for field in builtin_fields:
+            if field in inherit_flag_mapping:
+                inherit_flag = f"is_inherited_{inherit_flag_mapping[field]}"
+                queries.append(
+                    Q(**{inherit_flag: False, f"custom_{field}__icontains": keyword})
+                    | Q(**{inherit_flag: True, f"data_source_user__{field}__icontains": keyword})
+                )
+            else:
+                queries.append(Q(**{f"data_source_user__{field}__icontains": keyword}))
+
+        return queries
+
+    @staticmethod
+    def build_default_preview_tenant_user(tenant_id: str) -> TenantUser:
+        return TenantUser(
+            id="517hMkqnSBqF9Mv9",
+            data_source=DataSource(owner_tenant_id=tenant_id),
+            tenant_id=tenant_id,
+            data_source_user=DataSourceUser(
+                username="zhangsan",
+                full_name="张三",
+                phone="13512345671",
+                phone_country_code="86",
+                email="zhangsan@m.com",
+            ),
+        )
