+++ src/bk-user/bkuser/settings.py
@@ -787,8 +787,8 @@ ENABLE_COLLABORATION_TENANT = env.bool("ENABLE_COLLABORATION_TENANT", default=Fa
 INITIAL_ADMIN_USERNAME = env.str("INITIAL_ADMIN_USERNAME", "admin")
 # 内置租户管理员密码
 INITIAL_ADMIN_PASSWORD = env.str("INITIAL_ADMIN_PASSWORD", "")
-# 是否启用多数据源模式
-ENABLE_MULTI_DATA_SOURCE = env.bool("ENABLE_MULTI_DATA_SOURCE", default=False)
+# 是否启用多实体数据源模式
+ENABLE_MULTI_REAL_DATA_SOURCE = env.bool("ENABLE_MULTI_REAL_DATA_SOURCE", default=False)
 
 # logo 文件大小限制，单位为：KB
 MAX_LOGO_SIZE = env.int("MAX_LOGO_SIZE", 256)
